engine: seccomp based sandbox for AI runners.
The AIs are currently already isolated by the ruby VM, and should in
theory not be able to break out of it. However, it has a very large
attack surface, and has probably not been audited with untrusted code in
mind.

This sandbox provides a new layer of isolation, using Linux’s seccomp
mechanism to filter system calls the runner processes are allowed to
use.

It uses a whitelist of system calls, where only a small absolutely
required set of calls are allowed.

Once the sandbox is entered, the process may not open any new files or
sockets. However, previously opened files and sockets are still usable
as before. Thus, any resource required must be opened before entering
the sandbox and inherited by the runner process.

It supports three modes, kill, error and trace, which change the
behavior of the sandbox when a non-whitelisted system call is issued :
- Kill terminates the AI runner process.
- Error causes this specific system call to fail with a EPERM error.
- Trace allows the system call to run, but logs a message with the
  system call number and arguments. This does not offer any protection,
  but instead helps developing the whitelist.

Heroku deploy is configured to use trace mode, until we are confident
with the whitelist, at which point it will be changed to either kill or
error.

Additionally, when entering the sandbox, all environment variables are
clear, ensuring secret configuration values (tokens, session encryption
key, database credentials, ...) are not leaked.

+++ server/runner/client.go
@@ -9,6 +9,7 @@ import (
 	"os/exec"
 	"runtime"
 
+	"github.com/egnwd/outgain/server/config"
 	"github.com/egnwd/outgain/server/protocol"
 )
 
@@ -19,27 +20,49 @@ type RunnerClient struct {
 
 // StartRunner creates a new process to execute the AI runner.
 // `code` is used as the AI's source.
-func StartRunner(code string) (client *RunnerClient, err error) {
+func StartRunner(config *config.Config, code string) (client *RunnerClient, err error) {
 	fds, err := unix.Socketpair(unix.AF_LOCAL, unix.SOCK_STREAM, 0)
 	if err != nil {
 		return nil, err
 	}
 
 	clientFile := os.NewFile(uintptr(fds[0]), "client")
-	serverFile := os.NewFile(uintptr(fds[1]), "server")
 	defer func() {
 		if err != nil {
 			clientFile.Close()
 		}
 	}()
+
+	serverFile := os.NewFile(uintptr(fds[1]), "server")
 	defer serverFile.Close()
 
+	// runner may be sandboxed and not able to access /dev/urandom
+	// We open it here and pass it a file descriptor to it
+	randomFile, err := os.Open("/dev/urandom")
+	if err != nil {
+		return nil, err
+	}
+	defer randomFile.Close()
+
 	cmd := &exec.Cmd{
-		Path:       os.Args[0],
-		Args:       []string{"ai-runner"},
-		ExtraFiles: []*os.File{serverFile},
+		ExtraFiles: []*os.File{serverFile, randomFile},
 		Stdout:     os.Stdout,
 		Stderr:     os.Stderr,
+	}
+
+	if config.SandboxMode != "" {
+		cmd.Path = config.SandboxBin
+		cmd.Args = []string{
+			config.SandboxBin,
+			config.SandboxMode,
+			os.Args[0],
+			"ai-runner",
+		}
+	} else {
+		cmd.Path = os.Args[0]
+		cmd.Args = []string{
+			"ai-runner",
+		}
 	}
 
 	stdin, err := cmd.StdinPipe()
@@ -75,13 +98,21 @@ func (client *RunnerClient) Tick(state protocol.WorldState) (TickResult, error) 
 
 // Close closes the connection to the AI runner
 func (client *RunnerClient) Close() {
-	if err := client.client.Close(); err != nil {
-		log.Print("Error closing runner: ", err)
+	if client.client != nil {
+		if err := client.client.Close(); err != nil {
+			log.Print("Error closing runner: ", err)
+		}
+		client.client = nil
 	}
-	if err := client.cmd.Process.Kill(); err != nil {
-		log.Print("Error killing runner: ", err)
+	if client.cmd != nil {
+		if err := client.cmd.Process.Kill(); err != nil {
+			log.Print("Error killing runner: ", err)
+		}
+		cmd := client.cmd
+		go func() {
+			cmd.Wait()
+		}()
+
+		client.cmd = nil
 	}
-	go func() {
-		client.cmd.Wait()
-	}()
 }
